requirements.txt
langchain
langchain-community
PyPDF2
python-dotenv
streamlit
faiss-cpu
langchain-google-genai
pandas
google-generativeai